requirements.txt
astroid==2.3.3
attrs==19.3.0
autopep8==1.4.4
entrypoints==0.3
invoke==1.3.0
isort==4.3.21
kafka-python==1.4.7
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.0.2
packaging==19.2
pluggy==0.13.1
psutil==5.6.7
psycopg2-binary==2.8.4
py==1.8.0
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.4.4
pyparsing==2.4.5
python-dotenv==0.10.3
six==1.13.0
typed-ast==1.4.0
wcwidth==0.1.7
wrapt==1.11.2
